+++ src/full/Agda/Interaction/Options/Types.hs
@@ -68,7 +68,6 @@ data CommandLineOptions = Options
   , optGenerateVimFile       :: Bool
   , optIgnoreInterfaces      :: Bool
   , optIgnoreAllInterfaces   :: Bool
-  , optLocalInterfaces       :: Bool
   , optPragmaOptions         :: PragmaOptions
   , optOnlyScopeChecking     :: Bool
       -- ^ Should the top-level module only be scope-checked, and not type-checked?